Split parser into multiple files
