+++ src/parser/dml.rs
@@ -1,0 +1,64 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#[cfg(not(feature = "std"))]
+use alloc::{
+    string::{String, ToString},
+    vec,
+    vec::Vec,
+};
+
+use crate::ast::*;
+use crate::parser::{Parser, ParserError};
+use crate::tokenizer::{Token, Whitespace};
+
+impl<'a> Parser<'a> {
+    // Parse a `var = expr` assignment, used in an `UPDATE` statement
+    pub fn parse_assignment(&mut self) -> Result<Assignment, ParserError> {
+        let id = self.parse_identifier()?;
+        self.expect_token(&Token::Eq)?;
+        let value = self.parse_expr()?;
+        Ok(Assignment { id, value })
+    }
+
+    // Parse a tab separated values in COPY payload, used in a `COPY` statement
+    pub(super) fn parse_tab_values(&mut self) -> Vec<Option<String>> {
+        let mut values = vec![];
+        let mut content = String::from("");
+        while let Some(t) = self.next_token_no_skip() {
+            match t {
+                Token::Whitespace(Whitespace::Tab) => {
+                    values.push(Some(content.to_string()));
+                    content.clear();
+                }
+                Token::Whitespace(Whitespace::Newline) => {
+                    values.push(Some(content.to_string()));
+                    content.clear();
+                }
+                Token::Backslash => {
+                    if self.consume_token(&Token::Period) {
+                        return values;
+                    }
+                    if let Token::Word(w) = self.next_token() {
+                        if w.value == "N" {
+                            values.push(None);
+                        }
+                    }
+                }
+                _ => {
+                    content.push_str(&t.to_string());
+                }
+            }
+        }
+        values
+    }
+}
