+++ src/parser/tcl.rs
@@ -1,0 +1,68 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#[cfg(not(feature = "std"))]
+use alloc::{vec, vec::Vec};
+
+use crate::ast::*;
+use crate::dialect::keywords::Keyword;
+use crate::parser::{Parser, ParserError};
+use crate::tokenizer::Token;
+
+impl<'a> Parser<'a> {
+    pub fn parse_transaction_modes(&mut self) -> Result<Vec<TransactionMode>, ParserError> {
+        let mut modes = vec![];
+        let mut required = false;
+        loop {
+            let mode = if self.parse_keywords(&[Keyword::ISOLATION, Keyword::LEVEL]) {
+                let iso_level = if self.parse_keywords(&[Keyword::READ, Keyword::UNCOMMITTED]) {
+                    TransactionIsolationLevel::ReadUncommitted
+                } else if self.parse_keywords(&[Keyword::READ, Keyword::COMMITTED]) {
+                    TransactionIsolationLevel::ReadCommitted
+                } else if self.parse_keywords(&[Keyword::REPEATABLE, Keyword::READ]) {
+                    TransactionIsolationLevel::RepeatableRead
+                } else if self.parse_keyword(Keyword::SERIALIZABLE) {
+                    TransactionIsolationLevel::Serializable
+                } else {
+                    self.expected("isolation level", self.peek_token())?
+                };
+                TransactionMode::IsolationLevel(iso_level)
+            } else if self.parse_keywords(&[Keyword::READ, Keyword::ONLY]) {
+                TransactionMode::AccessMode(TransactionAccessMode::ReadOnly)
+            } else if self.parse_keywords(&[Keyword::READ, Keyword::WRITE]) {
+                TransactionMode::AccessMode(TransactionAccessMode::ReadWrite)
+            } else if required {
+                self.expected("transaction mode", self.peek_token())?
+            } else {
+                break;
+            };
+            modes.push(mode);
+            // ANSI requires a comma after each transaction mode, but
+            // PostgreSQL, for historical reasons, does not. We follow
+            // PostgreSQL in making the comma optional, since that is strictly
+            // more general.
+            required = self.consume_token(&Token::Comma);
+        }
+        Ok(modes)
+    }
+
+    pub fn parse_commit_rollback_chain(&mut self) -> Result<bool, ParserError> {
+        let _ = self.parse_one_of_keywords(&[Keyword::TRANSACTION, Keyword::WORK]);
+        if self.parse_keyword(Keyword::AND) {
+            let chain = !self.parse_keyword(Keyword::NO);
+            self.expect_keyword(Keyword::CHAIN)?;
+            Ok(chain)
+        } else {
+            Ok(false)
+        }
+    }
+}
